requirements.txt
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
sentry-sdk==1.16.0
drf-spectacular